Remove the college fields. Also fix a bad end bold tag.
There are two college fields which are hidden,
they must be initialised to a real value.
As they are hidden, a proper value will never be set through the UI.
Note also these fields are not used as the test in
_student_check_course allways returns TRUE!
So I have removed the fields and the call to _student_check_course.