student profile cck definitions
In this commit, also added support for installation
of the content type when the module is installed and
the removal of the content type when it gets uninstalled.

The student.install file was also modified in respect to
schemas which were removed and moved to the main veritas
(which should be called eduerp now) module.