Create views to generate a dummy staff and grading path to get staff/xxx menus to work
For some reason when installed, student/xxx, staff/xxx and grading/xxx menus
do not apear in the menus unless there is a student, staff and grading path
in existance. I do not understand this issue. But this does seem to fix it.
Note the menus do exist at a depth of 2 with a strange parent, but do not display.
Single level menus do not seem to work either!